Remove Obsolete members from EF 8
Fixes #34467
