+++ src/EFCore.Relational/Metadata/Conventions/RelationalQueryFilterRewritingConvention.cs
@@ -39,14 +39,6 @@ public class RelationalQueryFilterRewritingConvention : QueryFilterRewritingConv
             {
                 entityType.SetQueryFilter((LambdaExpression)DbSetAccessRewriter.Rewrite(modelBuilder.Metadata, queryFilter));
             }
-
-#pragma warning disable CS0618 // Type or member is obsolete
-            var definingQuery = ((IEntityType)entityType).GetDefiningQuery();
-            if (definingQuery != null)
-            {
-                entityType.SetDefiningQuery((LambdaExpression)DbSetAccessRewriter.Rewrite(modelBuilder.Metadata, definingQuery));
-            }
-#pragma warning restore CS0618 // Type or member is obsolete
         }
     }
