+++ src/EFCore/Metadata/Internal/CoreAnnotationNames.cs
@@ -201,15 +201,6 @@ public static class CoreAnnotationNames
     ///     any release. You should only use it directly in your code with extreme caution and knowing that
     ///     doing so can result in application failures when updating to a new Entity Framework Core release.
     /// </summary>
-    [Obsolete] // Remove with defining query
-    public const string DefiningQuery = "DefiningQuery";
-
-    /// <summary>
-    ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-    ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-    ///     any release. You should only use it directly in your code with extreme caution and knowing that
-    ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-    /// </summary>
     public const string EagerLoaded = "EagerLoaded";
 
     /// <summary>
@@ -411,9 +402,6 @@ public static class CoreAnnotationNames
         AfterSaveBehavior,
         BeforeSaveBehavior,
         QueryFilter,
-#pragma warning disable CS0612 // Type or member is obsolete
-        DefiningQuery,
-#pragma warning restore CS0612 // Type or member is obsolete
         EagerLoaded,
         LazyLoadingEnabled,
         ProviderClrType,
